mess_management/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-telegram-bot==20.7
qrcode==7.4.2
Pillow==10.1.0
cloudinary==1.36.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
cryptography==41.0.7
pytz==2023.3
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
